Move files to support a single project with multiple modules
When I originally built this project I had thought it would be a "base"
and "final" project set, like the other codelabs I'd been introduced to.
Having seen how to do a codelab with multiple modules, which greatly
improves maintenance and ease of understanding, I'm moving the files
around here to support this.

There are no code changes associated with this change.

Change-Id: I773e14323bb8b872dcc0839743cf31cef0c88bd1